qtrl-pro-project/requirements.txt
# Core Data & Computation
numpy==1.21.6
pandas==1.5.3
scipy==1.10.1

# Visualization
plotly==5.18.0
streamlit==1.36.0
matplotlib==3.7.5

# Machine Learning
tensorflow==2.12.0
keras==2.12.0
scikit-learn==1.2.2
shap==0.42.1
mlflow==2.5.0

# Quantum Computing
qiskit==1.0.0
qiskit-aer==0.13.0
qiskit-terra==1.0.0
qiskit-machine-learning==0.6.0
symengine==0.11.0

# Topological Data Analysis
giotto-tda==0.6.2
persim==0.3.5

# Financial Libraries
yfinance==0.2.28
pandas-ta==0.3.14b0
arch==6.3.0
pandas-market-calendars==4.4.0
exchange_calendars==4.11

# Trading API
alpaca-py==0.28.1

# Utilities
redis==5.0.1
joblib==1.3.2
ml-dtypes==0.2.0
protobuf==4.25.3